# docs/10-client-api/Client-API.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EDUagent Client-API",
    "version": "5.0.0",
    "description": "前端 (Vue3) ↔ Backend (FastAPI, Port 8001) 接口规范。\n统一返回格式: {code, message, data}。\n流式接口使用 SSE (text/event-stream)。\n前端发起刷新/生成接口后收到 task_id，并通过 /tasks/{task_id} 轮询；Backend 负责创建任务、记录归属用户和任务类型；Agent Service 长任务完成后由 Backend 接收 webhook 回调，前端不直接等待 webhook。"
  },
  "servers": [
    {
      "url": "http://localhost:8001/api/v1",
      "description": "本地开发环境"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/courses": {
      "get": {
        "tags": [
          "Courses"
        ],
        "summary": "我的课程列表",
        "description": "教师返回自己开的班，学生返回已加入的班。前端课程页/课程按钮使用此接口渲染；用户点击某门课程后，将 courses[].id 作为后续接口的 course_id；教师视角中 class_id 与 courses[].id 同源。v1 中 course_id 表示教师创建的课程/班级空间 ID，不拆分课程模板与教学班。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "courses": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CourseItem"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Courses"
        ],
        "summary": "创建课程/开班",
        "description": "仅教师可操作，返回唯一课程码 course_code 供学生加入。课程知识库由开发者预置并完成向量化，前端不上传课程原始资料；Backend 可为课程默认绑定预置知识库。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "created"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CreateCourseData"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "仅教师可创建",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "课程资源库不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "课程资源库未就绪",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/courses/join": {
      "post": {
        "tags": [
          "Courses"
        ],
        "summary": "通过课程码加入课程",
        "description": "仅学生可操作。course_code 是教师创建课程后分发给学生的课程码，不同于账号注册用 registration_code。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinCourseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "加入成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/JoinCourseData"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "仅学生可加入",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "课程码不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "已加入该课程",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/teaching/classes/{class_id}/students/{student_id}/learning": {
      "get": {
        "tags": [
          "Teaching"
        ],
        "summary": "查看学生学习情况汇总",
        "description": "教师只读查看学生在该课程内的学习看板，不触发学生画像、评估或学习路径刷新。该接口聚合学生端课程详情页核心展示数据。",
        "parameters": [
          {
            "name": "class_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "班级/课程 ID，与 /courses 返回的 courses[].id 相同"
          },
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/StudentLearning"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未登录或角色不符"
          },
          "403": {
            "description": "非该课程教师，无权查看"
          },
          "404": {
            "description": "课程不存在或学生未入班"
          }
        }
      }
    },
...
    "/evaluation": {
      "get": {
        "tags": [
          "Evaluation"
        ],
        "summary": "获取学习效果评估",
        "description": "只读取当前用户在该课程下最近一次学习效果评估，不实时调用 Agent。若尚无足够学习数据或尚未生成评估，返回空表格、空总结和 generated_at: null。",
        "parameters": [
          {
            "name": "course_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "课程 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/EvaluationData"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/evaluation/refresh": {
      "post": {
        "tags": [
          "Evaluation"
        ],
        "summary": "请求刷新学习效果评估",
        "description": "返回 task_id，异步处理。Backend 聚合练习记录、资源使用、学习路径进度、辅导对话摘要等 SQL 数据后调用 Agent 生成评估；完成后保存为该用户在该课程下的最近评估。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "course_id"
                ],
                "properties": {
                  "course_id": {
                    "type": "string",
                    "description": "课程 ID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "已接受",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 202
                    },
                    "message": {
                      "type": "string",
                      "example": "accepted"
                    },
                    "data": {
                      "$ref": "#/components/schemas/TaskResponse"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/learning-path": {
      "get": {
        "tags": [
          "LearningPath"
        ],
        "summary": "获取学习路径",
        "description": "只读取当前用户在该课程下最近一次学习路径，不实时调用 Agent。若尚未生成路径，返回 nodes: []、edges: []、current_position: null、generated_at: null。",
        "parameters": [
          {
            "name": "course_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/LearningPathData"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/learning-path/nodes/{node_id}/resources": {
      "get": {
        "tags": [
          "LearningPath"
        ],
        "summary": "路径节点资源推送",
        "description": "根据当前路径节点关联的知识点查询资源和练习。若资源不足，返回空数组，不视为错误。",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "course_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "课程ID，用于限定资源查询范围"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/NodeResources"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未登录"
          },
          "403": {
            "description": "未加入该课程"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CourseItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "课程 ID"
          },
          "name": {
            "type": "string",
            "description": "课程名称"
          },
          "description": {
            "type": "string",
            "description": "课程描述"
          },
          "course_code": {
            "type": "string",
            "description": "课程码"
          },
          "teacher_name": {
            "type": "string",
            "description": "教师姓名"
          },
          "student_count": {
            "type": "integer",
            "description": "学生人数"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "catalog_id": {
            "type": "string",
            "description": "绑定的共享课程资源库 ID；旧课程为空字符串"
          },
          "catalog_title": {
            "type": "string",
            "description": "绑定的共享课程资源库名称；旧课程为空字符串"
          }
        }
      },
      "CreateCourseData": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "课程 ID"
          },
          "name": {
            "type": "string",
            "description": "课程名称"
          },
          "course_code": {
            "type": "string",
            "description": "系统自动生成的唯一课程码"
          },
          "catalog_id": {
            "type": "string",
            "description": "绑定的共享课程资源库 ID；旧课程为空字符串"
          },
          "catalog_title": {
            "type": "string",
            "description": "绑定的共享课程资源库名称；旧课程为空字符串"
          }
        }
      },
      "CreateCourseRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "课程名称"
          },
          "description": {
            "type": "string",
            "description": "课程描述"
          },
          "catalog_id": {
            "type": "string",
            "description": "新教学班绑定的共享课程资源库 ID；Phase A 中可选，旧 courses 兼容路径可不传"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "错误码"
          },
          "message": {
            "type": "string",
            "description": "错误信息"
          },
          "data": {
            "type": "null"
          }
        }
      },
      "HTTPErrorResponse": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "object",
            "properties": {
              "code": {
                "type": "integer",
                "description": "错误码"
              },
              "message": {
                "type": "string",
                "description": "错误信息"
              },
              "data": {
                "type": "object",
                "nullable": true,
                "description": "固定为 null",
                "example": null
              }
            }
          }
        }
      },
      "EvaluationData": {
        "type": "object",
        "description": "用户在某门课程下的最近学习效果评估；尚未生成时返回空表格、空总结和 generated_at: null。",
        "properties": {
          "course_id": {
            "type": "string",
            "description": "课程 ID"
          },
          "progress_table": {
            "$ref": "#/components/schemas/TableData"
          },
          "mastery_table": {
            "$ref": "#/components/schemas/TableData"
          },
          "resource_usage_table": {
            "$ref": "#/components/schemas/TableData"
          },
          "node_progress": {
            "type": "array",
            "description": "按 KG 节点聚合的学习效果进度；无题节点使用 unassessed_default_pass。",
            "items": {
              "$ref": "#/components/schemas/EvaluationNodeProgress"
            }
          },
          "summary_text": {
            "type": "string",
            "description": "LLM 文字总结"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "生成时间；尚未生成时为 null"
          }
        }
      },
...
      "JoinCourseData": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "课程 ID"
          },
          "name": {
            "type": "string",
            "description": "课程名称"
          },
          "teacher_name": {
            "type": "string",
            "description": "教师姓名"
          }
        }
      },
      "JoinCourseRequest": {
        "type": "object",
        "required": [
          "course_code"
        ],
        "properties": {
          "course_code": {
            "type": "string",
            "description": "课程码"
          }
        }
      },
...
      "LearningPathData": {
        "type": "object",
        "description": "用户在某门课程下的最近学习路径；尚未生成时返回 nodes: []、edges: []、current_position: null、generated_at: null。",
        "properties": {
          "course_id": {
            "type": "string",
            "description": "课程 ID"
          },
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PathNode"
            }
          },
          "edges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PathEdge"
            }
          },
          "current_position": {
            "type": "object",
            "nullable": true,
            "properties": {
              "node_id": {
                "type": "string",
                "description": "当前节点 ID"
              },
              "node_name": {
                "type": "string",
                "description": "当前节点名称"
              }
            }
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "生成时间；尚未生成时为 null"
          }
        }
      },
...
      "PathEdge": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "description": "前置节点 ID"
          },
          "to": {
            "type": "string",
            "description": "后置节点 ID"
          }
        }
      },
...
      "StudentLearning": {
        "type": "object",
        "properties": {
          "student": {
            "type": "object",
            "description": "学生基础信息",
            "properties": {
              "id": {
                "type": "string",
                "description": "学生用户 ID"
              },
              "real_name": {
                "type": "string",
                "description": "真实姓名"
              },
              "student_id": {
                "type": "string",
                "description": "学号"
              }
            }
          },
          "evaluation_summary": {
            "type": "object",
            "properties": {
              "overall_score": {
                "type": "number",
                "nullable": true,
                "description": "最新评估知识点 average_score 的有效平均值 0-100；无有效掌握度数据时为 null"
              },
              "generated_at": {
                "type": "string",
                "format": "date-time",
                "description": "评估生成时间"
              },
              "summary_text": {
                "type": "string",
                "nullable": true,
                "description": "AI 评估摘要文本，由 Agent 生成"
              }
            }
          },
          "profile_summary": {
            "type": "object",
            "properties": {
              "knowledge_mastered": {
                "type": "integer",
                "description": "已掌握知识点数"
              },
              "knowledge_weak": {
                "type": "integer",
                "description": "薄弱知识点数"
              },
              "modal_preference": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "模态偏好排行"
              },
              "knowledge_coordinates": {
                "type": "array",
                "description": "知识坐标列表，来自用户画像",
                "items": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "知识点名称"
                    },
                    "status": {
                      "type": "string",
                      "description": "掌握状态：mastered / learning"
                    },
                    "mastered_at": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true,
                      "description": "掌握时间"
                    }
                  }
                }
              }
            }
          },
          "path_progress": {
            "type": "object",
            "properties": {
              "current_node": {
                "type": "string",
                "description": "当前学习节点名称"
              },
              "completed_nodes": {
                "type": "integer",
                "description": "已完成节点数"
              },
              "total_nodes": {
                "type": "integer",
                "description": "总节点数"
              }
            }
          },
          "quiz_stats": {
            "type": "object",
            "properties": {
              "total_attempts": {
                "type": "integer",
                "description": "总练习次数"
              },
              "avg_score": {
                "type": "number",
                "description": "平均正确率 0-100"
              },
              "avg_time_spent": {
                "type": "integer",
                "description": "平均耗时（秒）"
              },
              "mastery_breakdown": {
                "type": "array",
                "description": "按知识点拆分的练习正确率",
                "items": {
                  "type": "object",
                  "properties": {
                    "knowledge_point": {
                      "type": "string",
                      "description": "知识点名称"
                    },
                    "accuracy": {
                      "type": "number",
                      "description": "正确率 0-100"
                    }
                  }
                }
              }
            }
          },
          "weak_points": {
            "type": "array",
            "description": "按知识点聚合的错题 top 5，按 error_rate DESC, error_count DESC 排序；无错题时为空数组",
            "items": {
              "type": "object",
              "properties": {
                "knowledge_point": {
                  "type": "string",
                  "description": "知识点名称"
                },
                "error_count": {
                  "type": "integer",
                  "description": "错误次数"
                },
                "total_attempts": {
                  "type": "integer",
                  "description": "该知识点答题总数"
                },
                "error_rate": {
                  "type": "number",
                  "description": "错误率 0-1"
                }
              }
            }
          },
          "recent_activity": {
            "type": "array",
            "description": "最近 5 次 QuizSession，按 create_time DESC 排序；无记录时为空数组",
            "items": {
              "type": "object",
              "properties": {
                "quiz_id": {
                  "type": "string",
                  "description": "练习 ID"
                },
                "chapter": {
                  "type": "string",
                  "description": "章节"
                },
                "score": {
                  "type": "number",
                  "description": "得分（百分比）"
                },
                "correct_count": {
                  "type": "integer",
                  "description": "正确数"
                },
                "total_count": {
                  "type": "integer",
                  "description": "总题数"
                },
                "time_spent": {
                  "type": "integer",
                  "description": "耗时（秒）"
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time",
                  "description": "练习时间"
                }
              }
            }
          }
        }
      },
      "TableData": {
        "type": "object",
        "properties": {
          "columns": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "key": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                }
              }
            },
            "description": "表头"
          },
          "rows": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "数据行"
          }
        }
      },
      "TaskResponse": {
        "type": "object",
        "properties": {
          "task_id": {
            "type": "string",
            "description": "异步任务 ID"
          }
        }
      },
...
    }
  }
}